% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{add_ribbons_quantiles}
\alias{add_ribbons_quantiles}
\title{Helper function to add ribbons for forecasts.}
\usage{
add_ribbons_quantiles(g, qlist, k, col.fcst, alpha.ribbon)
}
\arguments{
\item{g}{}

\item{k}{}

\item{col.fcst}{}

\item{alpha.ribbon}{}

\item{z}{}
}
\value{
A ggplot object
}
\description{
Helper function to add ribbons for forecasts.
}
\keyword{internal}
