% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{aggregate_fcst}
\alias{aggregate_fcst}
\title{Helper function to aggregate forecasts}
\usage{
aggregate_fcst(var.to.aggregate, obj, simfwd)
}
\arguments{
\item{var.to.aggregate}{String. Name of the variable to aggregate.}

\item{obj}{REEM object.}

\item{simfwd}{Object of forward simulations, , as returned by the function \code{forecast()}.}
}
\value{
A list of dataframes.
}
\description{
Helper function to aggregate forecasts
}
