% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate values across time}
\usage{
aggregate_time(df, dt.aggr, var.name, aggreg.name = NULL)
}
\arguments{
\item{df}{Dataframe oftemporal data. 
Must have a variable named 
either \code{date} or \`code{time}.}

\item{dt.aggr}{Vector of integer (time) or dates (date)
were the aggregation is performed.}

\item{var.name}{String. Name of the variable to be aggregated.}
}
\value{
A dataframe where the values of the 
variable \code{var.name} has been aggregated (summed).
}
\description{
Aggregate values across time
}
\keyword{internal}
