% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calc_density_one}
\alias{calc_density_one}
\title{Calculate the density of forecasted values 
at a single date}
\usage{
calc_density_one(fcst.vals, density.n, density.adjust)
}
\arguments{
\item{fcst.vals}{}

\item{density.n}{}

\item{density.adjust}{}
}
\description{
Calculate the density of forecasted values 
at a single date
}
