% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{calc_score_one}
\alias{calc_score_one}
\title{Calculate the forecasting score at 
a given date of new observations}
\usage{
calc_score_one(
  i,
  var,
  obs.new,
  fcst,
  density.n,
  density.adjust,
  aggr.window,
  pb
)
}
\arguments{
\item{i}{}

\item{var}{}

\item{obs.new}{}

\item{fcst}{}

\item{density.n}{}

\item{density.adjust}{}

\item{aggr.window}{}
}
\description{
Calculate the forecasting score at 
a given date of new observations
}
