% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{check_date_start}
\alias{check_date_start}
\title{Check (and potentially set) the simulation start date}
\usage{
check_date_start(obj)
}
\arguments{
\item{obj}{}
}
\value{
Error message if appropriate.
}
\description{
Check (and potentially set) the simulation start date
}
