% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{err_fct}
\alias{err_fct}
\title{Error function for a fit algorithm}
\usage{
err_fct(
  cl.i,
  ww.i,
  obs.cl,
  obs.ww,
  use.cl,
  use.ww,
  err.type = "rel",
  do.plot = FALSE
)
}
\arguments{
\item{cl.i}{Dataframe of simulated clinical observations}

\item{ww.i}{Dataframe of simulated wastewater observations}

\item{obs.cl}{Dataframe of clinical observations}

\item{obs.ww}{Dataframe of wastewater observations}

\item{use.cl}{Logical. Use clinical data in fit?}

\item{use.ww}{Logical. Use wastewater data in fit?}
}
\value{
Numerical value representing the distance 
of the simulation from the observed data.
}
\description{
Error function for a fit algorithm
}
