% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_fit_aggreg}
\alias{extract_fit_aggreg}
\title{Extract posterior variable and aggregate}
\usage{
extract_fit_aggreg(obj, type, rename = TRUE)
}
\arguments{
\item{obj}{`reem` object}

\item{type}{String. Typcal values: `cl`, `ha`. (wastewater is NOT aggregated)}

\item{rename}{Logical. Rename aggregated variables.}
}
\value{
Dataframe.
}
\description{
Extract posterior variable and aggregate
}
\keyword{internal}
