% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{extract_helper}
\alias{extract_helper}
\title{Helper function that extract forecasted values
with the option to aggregate them (e.g. clinical cases)}
\usage{
extract_helper(i, fcst, aggr.window, obs.new, var)
}
\arguments{
\item{i}{}

\item{fcst}{}

\item{aggr.window}{}

\item{obs.new}{}

\item{var}{}
}
\value{
Numerical vector.
}
\description{
Helper function that extract forecasted values
with the option to aggregate them (e.g. clinical cases)
}
\keyword{internal}
