% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{get_psi}
\alias{get_psi}
\title{Title}
\usage{
get_psi(prm)
}
\arguments{
\item{prm}{List of distribution parameters}
}
\value{
Numerical vector representing  the distribution
of viral decay in wastewater.
}
\description{
Title
}
