% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{helper_aggreg}
\alias{helper_aggreg}
\title{Helper function to aggregate simulated observations}
\usage{
helper_aggreg(sim, type, dateobs, prms, var.name = "obs", percapita = FALSE)
}
\description{
Helper function to aggregate simulated observations
}
\keyword{internal}
