% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_prm_integer}
\alias{is_prm_integer}
\title{Check if the parameter should be an integer}
\usage{
is_prm_integer(prm.name)
}
\arguments{
\item{prm.name}{String. Name of the parameter}
}
\value{
Logical
}
\description{
Check if the parameter should be an integer
}
