% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_inc}
\alias{mean_inc}
\title{Mean incidence from the renewal equation.}
\usage{
mean_inc(t, R0, B, S, N, alpha, g, I)
}
\arguments{
\item{t}{Numerical. Time.}

\item{R0}{Numerical. Basic reproduction number.}

\item{B}{Numerical vector. Multiplicative factor changing the contact rate.
(if no change, set \code{B[t]=1} for all t)}

\item{S}{Numerical for the number of susceptible.}

\item{N}{Numerical. Total population size}

\item{alpha}{Numerical. Parameter representing mixing heterogeneity.
\code{alpha = 0} means homogeneous mixing.}

\item{g}{Numerical vector representing the intrinsic generation interval distribution.}

\item{I}{Numerical vector. Incidence at each time step.}
}
\description{
Mean incidence from the renewal equation.
}
