% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{normlarge}
\alias{normlarge}
\title{Normalization with one of the largest values}
\usage{
normlarge(x, largest.rank = 3)
}
\arguments{
\item{x}{Vector of positive numerical values.}
}
\value{
Normalized vector.
}
\description{
Normalization with one of the largest values
}
