% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem-tmp.R
\name{OLD_reem_simulate}
\alias{OLD_reem_simulate}
\title{Simulate an epidemic based on the renewal equation.}
\usage{
OLD_reem_simulate(prm, deterministic)
}
\arguments{
\item{prm}{List of model parameters.}

\item{deterministic}{Logical. Is the simulation deterministic.}
}
\description{
Simulate an epidemic base on a renewal equation that incorporates
both the clinical and wastewater data (ie fecal shedding is 
explicitly taken into account).
}
