% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_epi}
\alias{plot_epi}
\title{Plot simulated epidemic}
\usage{
plot_epi(simepi)
}
\arguments{
\item{simepi}{}
}
\value{
A list of two ggplot objects plotting
the time series of the population groups and
pathogen concentration in wastewater.
}
\description{
Plot simulated epidemic
}
