% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{plot_fitfcst}
\alias{plot_fitfcst}
\title{Helper function}
\usage{
plot_fitfcst(
  traj.fit,
  traj.fcst,
  obs,
  alpha.ribbon,
  col.fit,
  col.fcst,
  fcst.prm,
  title,
  ylab,
  qlist,
  xaxis
)
}
\arguments{
\item{traj.fit}{}

\item{traj.fcst}{}

\item{obs}{}

\item{alpha.ribbon}{}

\item{col.fit}{}

\item{col.fcst}{}

\item{fcst.prm}{}

\item{title}{}

\item{ylab}{}

\item{qlist}{}
}
\value{
a ggplot object
}
\description{
Helper function
}
\keyword{internal}
