% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{plot_forecast_scores}
\alias{plot_forecast_scores}
\title{Title}
\usage{
plot_forecast_scores(obj, obs.new, var, aggr.window = NULL)
}
\arguments{
\item{obj}{}

\item{obs.new}{}

\item{var}{}

\item{aggr.window}{}
}
\value{
A ggplot.
}
\description{
Title
}
