% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot observed data associated with a `reem` object.}
\usage{
plot_obs(obj)
}
\arguments{
\item{obj}{A `reem` object.}
}
\value{
A ggplot object.
}
\description{
Plot observed data associated with a `reem` object.
}
