% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prm_model_example}
\alias{prm_model_example}
\title{Returns an example of a list of model parameters}
\usage{
prm_model_example()
}
\value{
List of REEM model parameters
}
\description{
Returns an example of a list of model parameters
}
