% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem.R
\docType{class}
\name{reem-class}
\alias{reem-class}
\title{Class to represent a Renewal Equation Epidemic Model.}
\description{
Class to represent a Renewal Equation Epidemic Model.
}
\section{Fields}{

\describe{
\item{\code{name}}{character.}

\item{\code{prms}}{list.}

\item{\code{is.fitted}}{logical.}
}}


