% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reem_calc_scores}
\alias{reem_calc_scores}
\title{Forecasting scores given new observations}
\usage{
reem_calc_scores(var, obs.new, fcst, density.n = 100, density.adjust = 0.4)
}
\arguments{
\item{var}{}

\item{obs.new}{}

\item{fcst}{}

\item{density.n}{}

\item{density.adjust}{}
}
\description{
Forecasting scores given new observations
}
