% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_forecast}
\alias{reem_forecast}
\title{Forecast a fitted epidemic}
\usage{
reem_forecast(obj, prm.fcst, verbose)
}
\arguments{
\item{obj}{REEM object.}

\item{prm.fcst}{List. Forecast parameters.}

\item{verbose}{Logical. If \code{TRUE}, prints details.}
}
\value{
List containing forecast objects, 
with the following elements:
\item{asof: }{date when the forecast is done.}
\item{simfwd: }{object containing the forecasted trajectories}
}
\description{
Forecast a fitted epidemic
}
