% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reem_forecast_densities}
\alias{reem_forecast_densities}
\title{Densities of the forecasted values
given new observations.}
\usage{
reem_forecast_densities(
  var,
  obs.new,
  fcst,
  density.n,
  density.adjust,
  aggr.window
)
}
\arguments{
\item{var}{}

\item{obs.new}{}

\item{fcst}{}

\item{density.n}{}

\item{density.adjust}{}

\item{aggr.window}{}
}
\value{
List of dataframes. Each element is associated with an observation.
}
\description{
Densities of the forecasted values
given new observations.
}
