% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_forecast_peak}
\alias{reem_forecast_peak}
\title{Extract the forecasted peak values and timing.}
\usage{
reem_forecast_peak(var, fcst, obs)
}
\arguments{
\item{var}{String. Name of the model variable.}

\item{fcst}{Forecast object `fcst.obj` of the `reem` class object.}

\item{obs}{Dataframe. Observations associated with `var`.}
}
\value{
A dataframe of the forecasted peaks.
}
\description{
Extract the forecasted peak values and timing.
}
