% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_get_fcst_density}
\alias{reem_get_fcst_density}
\title{Retrieve the forecast posterior distribution}
\usage{
reem_get_fcst_density(
  var,
  date.future,
  fcst,
  density.n,
  density.adjust,
  aggr.window
)
}
\arguments{
\item{var}{String. Name of the variable forecasted. Typical choices are
`Y`, `Y.aggr` and `Wr`}

\item{date.future}{Vector of dates. Future dates where the variable `var` is forecasted.}

\item{fcst}{Forecast `reem` object.}

\item{density.n}{Integer. Number of equally spaced points at which 
the density is to be estimated. see \code{?density}.}

\item{density.adjust}{Numerical. Bandwidth adjustment. see \code{?density}.}

\item{aggr.window}{Integer or NULL. Time window to aggregate the values 
of the variable `var`. No aggregation if NULL.}
}
\value{
List of dataframes defining the densities as `y=f(x)`.
}
\description{
After the REEM model is fitted to the data,
epidemic trajectories are simulated by drawing the model 
parameters from their posterior distributions. This generates
"posterior trajectories" that are used for forecasting. 
This function returns, at given future time "slices", 
using the posterior trajectories, the forecasted (posterior) distribution
of the variable `var`.
}
\seealso{
\code{density()}
}
