% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{reem_inside_CI}
\alias{reem_inside_CI}
\title{Assess if the forecasted credible interval contains
the actual (future) observed value.}
\usage{
reem_inside_CI(var, obs.new, ci.width, aggr.window, fcst)
}
\arguments{
\item{var}{String. Name of the variable.}

\item{obs.new}{Dataframe of the new/actual observations of this variable.}

\item{ci.width}{Numerical. Width of the centered credible interval.}

\item{aggr.window}{Numerical. Width of the aggregation window for the variable. 
\code{NULL} if no aggregation.}

\item{fcst}{Forecast object of the \code{reem} object.}
}
\value{
A logical vector indicating if the actual observed value
is inside the credible interval. Each element corresponds to an 
observation date. (Hence the vector has the same length as the 
number of rows of \code{obs.new}.)
}
\description{
Assess if the forecasted credible interval contains
the actual (future) observed value.
}
