% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{reem_plot_fit}
\alias{reem_plot_fit}
\title{Plot fit outputs of a (fitted) reem object}
\usage{
reem_plot_fit(obj)
}
\arguments{
\item{obj}{fitted reem object}
}
\value{
List of ggplots.
}
\description{
Plot fit outputs of a (fitted) reem object
}
