% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem-func.R
\name{reem_plot_forecast}
\alias{reem_plot_forecast}
\title{Plot forecasts}
\usage{
reem_plot_forecast(obj, date_breaks, date_labels)
}
\arguments{
\item{obj}{}

\item{date_breaks}{}

\item{date_labels}{}
}
\value{
A ggplot object.
}
\description{
Plot forecasts
}
