% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_plot_peak}
\alias{reem_plot_peak}
\title{Plot timing and level of peak forecast}
\usage{
reem_plot_peak(var, obj, date_labels, logscale)
}
\arguments{
\item{var}{String. Name of the variable}

\item{obj}{List. Object}

\item{date_labels}{String. Date labels (as understood by \code{ggplot}).}

\item{logscale}{Logical. Use logarithmic scale for y axis. Default is \code{FALSE}.}
}
\value{
A ggplot object plotting the time and value of the forecasted peaks.
}
\description{
Plot timing and level of peak forecast
}
