% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_proba_box}
\alias{reem_proba_box}
\title{Returns the probability that the forecasted variable trajectory
is within a box defined by lower and upper dates and values.}
\usage{
reem_proba_box(var, date.lower, date.upper, val.lower, val.upper, fcst)
}
\arguments{
\item{var}{String. Name of the variable.}

\item{date.lower}{Date. Lower date defining the box.}

\item{date.upper}{Date. Upper date defining the box.}

\item{val.lower}{Numeric. Lower value defining the box.}

\item{val.upper}{Numeric. Upper value defining the box.}

\item{fcst}{Forecast object as returned by the function \code{forecast()}.}
}
\value{
Numeric. The corresponding probability.
}
\description{
Returns the probability that the forecasted variable trajectory
is within a box defined by lower and upper dates and values.
}
