% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem-func.R
\name{reem_simulate}
\alias{reem_simulate}
\title{Simulate an epidemic with a REEM.}
\usage{
reem_simulate(prms, deterministic)
}
\arguments{
\item{prms}{}

\item{deterministic}{}
}
\description{
Simulate an epidemic with a REEM.
}
