% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{reem_simulate_epi}
\alias{reem_simulate_epi}
\title{Simulate a full epidemic with a REEM
including clinical and wastewater "observations".}
\usage{
reem_simulate_epi(obj, deterministic)
}
\arguments{
\item{obj}{A object of class `reem`}

\item{deterministic}{Logical. Is the simulation deterministic? (or stochastic)}
}
\value{
A list
}
\description{
Simulate a full epidemic with a REEM
including clinical and wastewater "observations".
}
