% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{reem_traj_dist_obs}
\alias{reem_traj_dist_obs}
\title{Calculate the distance of the 
epidemic trajectory from 
clinical and wastewater observations
for a given set of REEM parameters.}
\usage{
reem_traj_dist_obs(
  obj,
  use.cl,
  use.ha,
  use.ww,
  err.type,
  deterministic,
  n.sim = 10,
  verbose = FALSE
)
}
\arguments{
\item{use.cl}{}

\item{use.ww}{}

\item{err.type}{}

\item{deterministic}{}

\item{n.sim}{}

\item{verbose}{}

\item{obs.cl}{}

\item{obs.ww}{}

\item{prms}{}
}
\value{
List containing:
\item{distance: }{Numerical. The distance ofthe simulation from observations}
\item{sim: }{Dataframe of the simulations.}
}
\description{
Calculate the distance of the 
epidemic trajectory from 
clinical and wastewater observations
for a given set of REEM parameters.
}
