\name{rif-package}
\alias{rif-package}
\alias{rif}
\docType{package}
\title{
\packageTitle{rif}
}
\description{
\packageDescription{rif}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rif}
\packageIndices{rif}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rif}

Maintainer: \packageMaintainer{rif}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
