% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_I_init}
\alias{set_I_init}
\title{Translate the initial PROPORTION 
of infected individuals into a
NuMBER of individuals initially infected.}
\usage{
set_I_init(prm)
}
\arguments{
\item{prm}{List. Model parameters.}
}
\description{
Translate the initial PROPORTION 
of infected individuals into a
NuMBER of individuals initially infected.
}
