% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{set_obs_schedule}
\alias{set_obs_schedule}
\title{Set the schedule of times and dates for the observations (if any).
Handles `date.start` changes.}
\usage{
set_obs_schedule(type, obj)
}
\arguments{
\item{type}{string. variable type}

\item{obj}{reem object}
}
\value{
a reem object
}
\description{
Set the schedule of times and dates for the observations (if any).
Handles `date.start` changes.
}
\keyword{internal}
