% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ssrs}
\alias{ssrs}
\title{sqrt sum of squared relative difference}
\usage{
ssrs(x, y)
}
\arguments{
\item{x}{simulations}

\item{y}{observations}
}
\value{
Numerical. The sum of squared relative difference
}
\description{
sqrt sum of squared relative difference
}
\keyword{internal}
