% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sss}
\alias{sss}
\title{sqrt sum of squared difference}
\usage{
sss(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
Numerical. The sum of squared difference
}
\description{
sqrt sum of squared difference
}
\keyword{internal}
