% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summarise_post_traj}
\alias{summarise_post_traj}
\title{Helper function to summarize 
NON-AGGREGATED posterior trajectories}
\usage{
summarise_post_traj(obj, obsdata, varname)
}
\description{
Helper function to summarize 
NON-AGGREGATED posterior trajectories
}
\keyword{internal}
