% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{summarize_fcst}
\alias{summarize_fcst}
\title{Calculate quantiles of forecast trajectories}
\usage{
summarize_fcst(simfwd, prm.fcst, vars = NULL)
}
\arguments{
\item{simfwd}{Object of forecast trajectories.}

\item{prm.fcst}{List of forecast parameters. 
In particular, must contain an element named `ci`
specifying the quantiles probabilities to calculate.}

\item{vars}{String vector. If not NULL (default), overrides `prm.fcst$vars.to.fcst`.}
}
\value{
Dataframe of quantiles.
}
\description{
Calculate quantiles of forecast trajectories
}
\keyword{internal}
