% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_Bt}
\alias{update_Bt}
\title{Update the values for B(t) based on 
prior or posterior samples.
This is used when the value of B(t > t*)
was fitted (for a given date t*) by 
specifying `Byyyymmdd` in the parameters
to fit (`prms.to.fit`) where t*=yyyymmdd.}
\usage{
update_Bt(obj, pp)
}
\arguments{
\item{obj}{reem object}

\item{pp}{List of the fitted parameters}
}
\value{
the same reem object `obj` but with Bt update accordingly.
}
\description{
Update the values for B(t) based on 
prior or posterior samples.
This is used when the value of B(t > t*)
was fitted (for a given date t*) by 
specifying `Byyyymmdd` in the parameters
to fit (`prms.to.fit`) where t*=yyyymmdd.
}
