% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{generate_priors}
\alias{generate_priors}
\title{Generate prior samples given 
paramter names and range.}
\usage{
generate_priors(prms.to.fit, n.priors)
}
\arguments{
\item{prms.to.fit}{List of parameters name with their 
prior distribution and parameters}

\item{n.priors}{Integer. Number of sample to generate for each parameter.}
}
\description{
Generate prior samples given 
paramter names and range.
}
