% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.R
\name{get_gi}
\alias{get_gi}
\title{Title}
\usage{
get_gi(prms = list(m = 3, v = 1, tmax = 10), type = "gamma")
}
\arguments{
\item{prms}{List of distribution parameters}

\item{type}{String. Distribution family (e.g., gamma)}
}
\value{
Numerical vector representing 
the intrinsic generation interval distribution.
}
\description{
Title
}
