% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem.R
\docType{class}
\name{reem-class}
\alias{reem-class}
\title{Class to represent a Renewal Equation Epidemic Model.}
\description{
Class to represent a Renewal Equation Epidemic Model.
}
\section{Fields}{

\describe{
\item{\code{name}}{Character for the name of the model.}

\item{\code{prms}}{List of model parameters.}

\item{\code{is.fitted}}{Logical. Is the model fitted to the observation
data attached (\code{obs.cl} and/or \code{obs.ww})?}

\item{\code{obs.cl}}{Dataframe representing the clinical observations. 
Must have \code{date} and \code{obs} variables.}

\item{\code{obs.ww}}{Dataframe representing the wastewater observations. 
Must have \code{date} and \code{obs} variables.}

\item{\code{fit.obj}}{List containing the fitted object.}

\item{\code{fit.prm}}{List that defines the parameters for the fitting algorithm.}
}}


