% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem.R
\docType{class}
\name{reem-class}
\alias{reem-class}
\title{Class to represent a Renewal Equation Epidemic Model.}
\value{
An object of class \code{reem}.
}
\description{
Class to represent a Renewal Equation Epidemic Model.
}
\section{Fields}{

\describe{
\item{\code{name}}{String for the name of the model. 
Just for convenience when working with multiple models,
this does not affect any calculation.}

\item{\code{prms}}{List of model parameters. (NEEDS TO BE CHECKED!)
\itemize{
  \item{\code{horizon}: }{horizon time of the simulation.}
  \item{\code{last.obs}: }{last observation time (must be < \code{horizon}).}
  \item{\code{B}: }{numerical vector representing the time 
  dependent multiplicative factor for transmission.}
  \item{\code{date.start}: }{start date of the epidemic.}
  \item{\code{R0}: }{basic reproduction number.}
  \item{\code{N}: }{population size.}
  \item{\code{alpha}: }{transmission heterogeneity (alpha=-Inf: homogeneous).}
  \item{\code{I.init}: }{initial incidence (overwritten in fit ABC).}
  \item{\code{lag}: }{aggregation window for clinical reports.}
  \item{\code{rho}: }{ mean reporting ratio.}
  \item{\code{g}: }{intrinsic generation interval distribution.}
  \item{\code{fec}: }{fecal shedding kinetics.}
  \item{\code{h.prop}: }{total proportion of infections that will be hospitalized.}
  \item{\code{h.lags}: }{time lags repartition weights between infection and hospital admission times.}
  \item{\code{kappa}: }{pathogen decay in wastewater.}
  \item{\code{psi}: }{plug flow simulation.}
  \item{\code{shed.mult}: }{deposited fecal shedding multiplier.}
}}

\item{\code{is.fitted}}{Logical. Is the model fitted to the observation
data attached (\code{obs.cl} and/or \code{obs.ww})?}

\item{\code{obs.cl}}{Dataframe representing the clinical observations. 
Must have \code{date} and \code{obs} variables.}

\item{\code{obs.ha}}{Dataframe representing the hospital admissions. 
Must have \code{date} and \code{obs} variables.}

\item{\code{obs.ww}}{Dataframe representing the wastewater observations. 
Must have \code{date} and \code{obs} variables.}

\item{\code{fit.obj}}{List containing the fitted object.}

\item{\code{fit.prm}}{List that defines the parameters for the fitting algorithm.}
}}


