% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{reem_fit_abc}
\alias{reem_fit_abc}
\title{Fit model to observed data using the ABC algorithm.}
\usage{
reem_fit_abc(obj, prm.abc, prms.to.fit, verbose = FALSE)
}
\arguments{
\item{obj}{}

\item{prm.abc}{}

\item{prms.to.fit}{}
}
\value{
List storing trajectories, ABC distances and posterior samples.
}
\description{
Fit model to observed data using the ABC algorithm.
}
