% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reem_proba_box}
\alias{reem_proba_box}
\title{Returns the probability that the forecasted variable
is within a box defined by lower and upper dates and values.}
\usage{
reem_proba_box(var, date.lower, date.upper, val.lower, val.upper, fcst)
}
\arguments{
\item{var}{}

\item{date.lower}{}

\item{date.upper}{}

\item{val.lower}{}

\item{val.upper}{}

\item{fcst}{}
}
\description{
Returns the probability that the forecasted variable
is within a box defined by lower and upper dates and values.
}
