% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reem-func.R
\name{reem_simulate_epi}
\alias{reem_simulate_epi}
\title{Simulate a full epidemic with a REEM
including clinical and wastewater "observations".}
\usage{
reem_simulate_epi(prms, deterministic)
}
\arguments{
\item{variables}{}
}
\description{
Simulate a full epidemic with a REEM
including clinical and wastewater "observations".
}
